test: split GridView into smaller test pages
